Added initial support for Hebrew, defined trees in code
This commit adds initial, partly broken, support for Hebrew
on the website. To make this easier, we include a definition
of the site tree in code rather than a fixture.

Note the change in the "operation instructions" reflecting
this -- we no longer load the sitetrees fixture, instead
we use the sutetree management command to load data from source.